Change filename to useMobile.js and add documentation
